It is the file where all the file of setmate app is  uploaded...
It is a project related to the college dating app. It is build using the xml for fronted and java language for backend , here we add firebase for the storage and user verification system. Also , firebase real time system is used here so, Real base chatting system is also applied in this platform.